python-service/requirements.txt
requests
python-dotenv
py-directus